refactor: GameResult 클래스 구조 개선
GameResult와 PlayerResult를 별도 파일로 분리
