+++ src/main/java/racingcar/service/GameService.java
@@ -2,10 +2,10 @@ package racingcar.service;
 
 import java.util.List;
 import racingcar.domain.game.Game;
-import racingcar.domain.game.GameResult;
 import racingcar.domain.movement.MovementPolicy;
 import racingcar.domain.player.Player;
 import racingcar.dto.request.GameStartRequest;
+import racingcar.dto.response.GameResult;
 import racingcar.dto.response.Winners;
 import racingcar.exception.game.GameException;
 
@@ -50,5 +50,5 @@ public class GameService {
     public Winners getWinners() {
         return Winners.from(game.getPlayers());
     }
-    
+
 }
