+++ src/test/java/racingcar/service/GameServiceTest.java
@@ -5,10 +5,10 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.game.GameResult;
 import racingcar.domain.movement.MovementPolicy;
 import racingcar.domain.movement.RandomMovementStrategy;
 import racingcar.dto.request.GameStartRequest;
+import racingcar.dto.response.GameResult;
 import racingcar.dto.response.Winners;
 import racingcar.exception.game.GameException.GameNotInitializedException;
 import racingcar.mock.AlternateNumberGenerator;
